backend/requirements.txt
# Core Web Framework
django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter

# Production Server & Database
gunicorn
whitenoise
dj-database-url
python-dotenv
psycopg2-binary

# AI & Third Party
google-generativeai
requests
twilio

# Utilities
Pillow
qrcode[pil]
PyPDF2
